+++ src/ua/data_types/browse_result.rs
@@ -4,6 +4,11 @@ crate::data_type!(BrowseResult);
 
 impl BrowseResult {
     #[must_use]
+    pub const fn status_code(&self) -> ua::StatusCode {
+        ua::StatusCode::new(self.0.statusCode)
+    }
+
+    #[must_use]
     pub fn references(&self) -> Option<ua::Array<ua::ReferenceDescription>> {
         // TODO: Adjust signature to return non-owned value instead.
         ua::Array::from_raw_parts(self.0.references, self.0.referencesSize)